Implement RCC commands and  live compilation for QRC file with RCC.
