+++ src/live-compilations.ts
@@ -1,7 +1,8 @@
 import { merge, Subscription } from 'rxjs';
-import { liveCompilation$ as liveUiCompilation$ } from './uic/live-compilation';
+import { liveCompilation$ as liveResourceCompilation$ } from './tools/rcc';
+import { liveCompilation$ as liveUiCompilation$ } from './tools/uic';
 
-const liveCompilations$ = merge(liveUiCompilation$);
+const liveCompilations$ = merge(liveUiCompilation$, liveResourceCompilation$);
 
 let subscription: Subscription | undefined;
