--- src/uic/commands.ts
@@ -1,13 +1,0 @@
-import * as vscode from 'vscode';
-import {
-  getActiveWorkspaceFolderPath,
-  getFsPathOrActiveDocumentPath,
-} from '../paths';
-import { compile } from './uic';
-
-export async function compileForm(fileUri?: vscode.Uri) {
-  return compile({
-    inPath: getFsPathOrActiveDocumentPath(fileUri),
-    cwd: getActiveWorkspaceFolderPath(),
-  });
-}
