+++ src/utils/predefined-variable-resolver.ts
